refactor(storage): split GrobStore into per-concern submodules
`storage/mod.rs` had grown to 841 LOC mixing four concerns in one impl.
Split the methods into cohesive submodules of cross-file `impl GrobStore`
blocks (child modules, so they keep access to the struct's private fields):

- `storage/spend.rs`         — load_spend / record_spend / flush_spend
- `storage/oauth.rs`         — OAuth token CRUD + list_enc_files helper
- `storage/secrets_store.rs` — named-secret CRUD (distinct from the
                                pluggable `secrets` backends module)
- `storage/vkeys.rs`         — virtual-key CRUD

`mod.rs` keeps the struct, `open`, `default_path`, `path`, the shared
`sanitize_filename` helper, and the tests. Public API is byte-for-byte
unchanged; behavior-preserving (all existing tests pass unmodified —
1193 lib + 267 integration). Closes the last audit maintainability item.
